+++ src/main/java/models/Venue.java
@@ -3,10 +3,10 @@ package models;
 public class Venue {
 
 	private String id;
-	private Long latitude;
-	private Long longitude;
+	private Double latitude;
+	private Double longitude;
 
-	public Venue(String id, Long latitude, Long longitude) {
+	public Venue(String id, Double latitude, Double longitude) {
 		this.id = id;
 		this.latitude = latitude;
 		this.longitude = longitude;
@@ -47,11 +47,11 @@ public class Venue {
 		return id;
 	}
 
-	public Long getLatitude() {
+	public Double getLatitude() {
 		return latitude;
 	}
 
-	public Long getLongitude() {
+	public Double getLongitude() {
 		return longitude;
 	}
 }
